src/main/java/cz/upce/fei/inptp/databasedependency/business/AuthenticationService.java
package cz.upce.fei.inptp.databasedependency.business;

import cz.upce.fei.inptp.databasedependency.dao.PersonDAO;
import cz.upce.fei.inptp.databasedependency.entity.Person;
import java.io.UnsupportedEncodingException;
import java.math.BigInteger;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.logging.Level;
import java.util.logging.Logger;

/**
 * Authentication service - used for authentication of users stored in db.
 * Authentication should success if login and password (hashed) matches.
 */
public class AuthenticationService {

    private PersonDAO personDao;

    public AuthenticationService(PersonDAO personDao) {
        this.personDao = personDao;
    }

    public boolean authenticate(String login, String password) {
        Person person = personDao.load("name = '" + login + "'");
        if (person == null) {
            return false;
        }

        return person.getPassword().equals(encryptPassword(password));
    }

    // TODO: (low priority) - change to safe password hash function (PBKDF2, bcrypt, scrypt)
    public static String encryptPassword(String password) {
        try {
            MessageDigest crypt = MessageDigest.getInstance("SHA-1");
            crypt.reset();
            crypt.update(password.getBytes("UTF-8"));

            return new BigInteger(1, crypt.digest()).toString(16);
        } catch (NoSuchAlgorithmException | UnsupportedEncodingException ex) {
            Logger.getLogger(AuthenticationService.class.getName()).log(Level.SEVERE, null, ex);
        }
        return null;
    }

}
